{"types":"C","runs":3}

+++ Problem023.java
@@ -2,7 +2,7 @@ import acm.program.ConsoleProgram;
 
 public class Problem023 extends ConsoleProgram {
 	public void run(){
-		int n = readInt("Enter n: ");
+		double n = readInt("Enter n: ");
 		int sum = calculateSum(n);
 		println("sum of number 1...n is: " + sum);
 		println(numberOfDivisors(n));
@@ -16,7 +16,7 @@ public class Problem023 extends ConsoleProgram {
 		return sum;
 	}
 	private int numberOfDivisors(int n){
-		int num = 0;
+		double num = 0;
 		for(int i = 0; i <= n; i++){
 			if (n % i == 0){
 				num += 1;
